+++ constants.py
@@ -1,0 +1,10 @@
+# directory naming
+UNLABELLED_FOLDER = 'unlabelled'
+LABELLED_FOLDER = 'labelled'
+
+# class names
+FOLDERS = ['-1', '0', '1', 'pending']
+
+# set constants for window size
+MAX_WINDOW_HEIGHT = 600
+MAX_WINDOW_WIDTH = 600
